# orca-backend/openapi2.json
{"openapi":"3.1.0","info":{"title":"ORCA Backend","description":"Marine EcoSystem Reasoning with Collaborative Agents","version":"0.1.0"},"paths":{"/api/v1/query/":{"post":{"tags":["query"],"summary":"Submit Query","description":"Legacy submission route","operationId":"submit_query_api_v1_query__post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Query Data"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/query/{query_id}":{"get":{"tags":["query"],"summary":"Get Query Result","description":"Legacy retrieval route","operationId":"get_query_result_api_v1_query__query_id__get","parameters":[{"name":"query_id","in":"path","required":true,"schema":{"type":"string","title":"Query Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/query":{"post":{"tags":["orca_api"],"summary":"Submit Api Query","description":"POST /api/query\nInput: raw user text (+ optional lat/lon if browser geolocation is available)\nOutput: { query_id: str, status: \"processing\" }","operationId":"submit_api_query_api_query_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryPostInput"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/result/{query_id}":{"get":{"tags":["orca_api"],"summary":"Get Api Result","description":"GET /api/result/{query_id}\nOutput: RiskCard (poll until status = \"done\"; return partial/loading state otherwise)","operationId":"get_api_result_api_result__query_id__get","parameters":[{"name":"query_id","in":"path","required":true,"schema":{"type":"string","title":"Query Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RiskCard"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/agents/status":{"get":{"tags":["orca_api"],"summary":"Get Api Agents Status","description":"GET /api/agents/status\nOutput: list[AgentStatus] - live health check of each specialist agent","operationId":"get_api_agents_status_api_agents_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AgentStatus"},"type":"array","title":"Response Get Api Agents Status Api Agents Status Get"}}}}}}},"/":{"get":{"summary":"Root","description":"Root endpoint - health check","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AgentStatus":{"properties":{"agent_name":{"type":"string","title":"Agent Name","description":"Name of the specialist agent"},"status":{"type":"string","title":"Status","description":"Status string: 'ok', 'degraded', or 'failed'"},"note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Note","description":"Optional notes on health status"}},"type":"object","required":["agent_name","status"],"title":"AgentStatus","description":"Status representation for specialist agents"},"EvidenceItem":{"properties":{"claim":{"type":"string","title":"Claim","description":"Statement claim"},"source":{"type":"string","title":"Source","description":"Sensor source name"},"field":{"type":"string","title":"Field","description":"Data field key"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp","description":"Observation timestamp"},"confidence":{"type":"number","title":"Confidence","description":"Sensor confidence level"}},"type":"object","required":["claim","source","field","timestamp","confidence"],"title":"EvidenceItem","description":"A claim traceable to a source telemetry data point"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"QueryPostInput":{"properties":{"text":{"type":"string","title":"Text"},"lat":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lat"},"lon":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lon"}},"type":"object","required":["text"],"title":"QueryPostInput","description":"Input payload for submitting a raw query"},"RiskCard":{"properties":{"risk_level":{"type":"string","title":"Risk Level","description":"Calculated threat rating: 'low', 'medium', or 'high'"},"reasoning":{"type":"string","title":"Reasoning","description":"Natural language explanation of risk"},"recommendation":{"type":"string","title":"Recommendation","description":"Safety actions and parameters recommendation"},"evidence":{"items":{"$ref":"#/components/schemas/EvidenceItem"},"type":"array","title":"Evidence","description":"Sensor data points backing the claim"},"agent_status":{"items":{"$ref":"#/components/schemas/AgentStatus"},"type":"array","title":"Agent Status","description":"Responding agent states"},"status":{"type":"string","title":"Status","description":"Parsing status: 'processing', 'done', or 'failed'","default":"done"}},"type":"object","required":["risk_level","reasoning","recommendation"],"title":"RiskCard","description":"The safety card produced by ORCA's synthesis engine"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}